+++ main.go
@@ -1,7 +1,7 @@
 package main
 
 import (
-	"os"
+	"log"
 
 	"github.com/generationtux/brizo/app"
 	"github.com/joho/godotenv"
@@ -11,6 +11,14 @@ func main() {
 	// .env file for local configuration during development (see .env.example)
 	godotenv.Load()
 
-	cli := app.ConfigureCLI()
-	cli.Run(os.Args)
+	brizo := app.New()
+
+	// Intializing Brizo will make sure everything is configured correctly
+	// and perform any migration tasks that are required.
+	err := brizo.Initialize()
+	if err != nil {
+		log.Fatalln(err)
+	}
+
+	// Run Brizo to handle the current command
 }
